requirements.txt
datasets==2.16.1
scikit-learn==1.3.1
scikit-learn-intelex==2021.20220215.212715
thefuzz==0.19.0
transformers==4.47.1
torch==2.5.1
serpapi==0.1.5
openai==1.58.1
google-generativeai==0.8.2
google-ai-generativelanguage==0.6.10